Move frontend files to src/
