+++ src/components/HomeContent/HomeContent.scss
