+++ src/components/NeighborIcon.js
