+++ src/components/layout/ComingSoon.js
