+++ src/utils/auth/firebaseSessionHandler.js
